backend/requirements.txt
Django==4.2
psycopg2-binary==2.9.6

redis==4.5.5
kafka-python==2.0.2

loguru==0.7.3
